Move InlinePanel include module to components
